In git-delete-branches.sh added error message if the current branch is in the list of branches to be deleted.

+++ Tools/my-tools/git-delete-branches.sh
@@ -5,6 +5,14 @@ if [ -z "$1" ]; then
 	exit 0
 fi
 
-for branch in $(git branch | grep "$1"); do
+BRANCHES=$(git branch | ag --case-sensitive -- "$1")
+CURRENT_BRANCH=$(git rev-parse --abbrev-ref HEAD)
+
+if [[ $BRANCHES == *"$CURRENT_BRANCH"* ]]; then
+	echo "Cannot delete checked out branch. Exiting."
+	exit 1
+fi
+
+for branch in $BRANCHES; do
 	git push --delete origin "$branch" && git branch -D "$branch"
 done
